Fix functions expecting interface{} that now receive QueryResults
